Update issue 6
remove all the *8 from real*8. Now there should only be 'real'. 
This should eliminate the problems around the conversion

+++ src/TEST1.f
@@ -8,9 +8,9 @@ C-----------------------------------------------------------------
 
       INTEGER NS,NC,NAR,ISEL,NREG,NLEG,NENERGY,NCOF,
      +        IC,NZM,I,J,NMAXGR,ISHOT
-      REAL*8 M,T,DEN,DS,CFF1,CFF2,CFF3,CFF4,XI,TAU,COEFF,
-     +       EPS, SIGMA, NORM, ZSP, EPARR
-      REAL*8 RHO, RN, E, Q, BN
+      REAL M,T,DEN,DS,CFF1,CFF2,CFF3,CFF4,XI,TAU,COEFF,
+     +      EPS, SIGMA, NORM, ZSP, EPARR
+      REAL RHO, RN, E, Q, BN
       LOGICAL NEOGEO, NEOFRC
 
       include 'elem_config.inc'
