+++ src/TEST4.f
@@ -9,10 +9,10 @@ C-----------------------------------------------------------------
 
       INTEGER NS,NC,NAR,ISEL,NREG,NLEG,NENERGY,NCOF,
      +        IC,NZM,I,J,NMAXGR,K,L,ISHOT
-      REAL*8 M,T,DEN,DS,CFF1,CFF2,CFF3,CFF4,XI,TAU,C1,C2,C3,
+      REAL M,T,DEN,DS,CFF1,CFF2,CFF3,CFF4,XI,TAU,C1,C2,C3,
      +       EPS, SIGMA, NORM, ALPHA, OMTH, OMTI, RESUL,
      +       RESUA,ZSP,EPARR
-      REAL*8 RHO, RN, E, Q, BN
+      REAL RHO, RN, E, Q, BN
       LOGICAL NEOGEO, NEOFRC
 
       include 'elem_config.inc'
