+++ src/TEST6.f
@@ -9,8 +9,8 @@ C-----------------------------------------------------------------
 
       INTEGER NS,NC,NAR,ISEL,NREG,NLEG,NENERGY,NCOF,
      +        IC,NZM,I,J,NMAXGR,L
-      REAL*8 M,T,DEN,DS,CFF4,XI,TAU, EPS, SIGMA, NORM, ZSP
-      REAL*8 RHO, RN, E, Q, BN
+      REAL M,T,DEN,DS,CFF4,XI,TAU, EPS, SIGMA, NORM, ZSP
+      REAL RHO, RN, E, Q, BN
       LOGICAL NEOGEO, NEOFRC
 
       PARAMETER(NAR = 5)
